Conflicts:
	configure
	mkspecs/common/symbian/symbian.conf
	mkspecs/features/symbian/application_icon.prf
	qmake/generators/makefile.cpp
	qmake/generators/symbian/initprojectdeploy_symbian.cpp
	qmake/generators/symbian/symmake.cpp
	tools/assistant/tools/assistant/assistant.pro

+++ src/opengl/gl2paintengineex/qgl2pexvertexarray.cpp
@@ -145,7 +145,7 @@ void QGL2PEXVertexArray::addPath(const QVectorPath &path, GLfloat curveInverseSc
                 // threshold based on same algorithm as in qtriangulatingstroker.cpp
                 int threshold = qMin<float>(64, qMax(bounds.width(), bounds.height()) * 3.14f / (curveInverseScale * 6));
                 if (threshold < 3) threshold = 3;
-                qreal one_over_threshold_minus_1 = 1.f / (threshold - 1);
+                qreal one_over_threshold_minus_1 = qreal(1) / (threshold - 1);
                 for (int t=0; t<threshold; ++t) {
                     QPointF pt = b.pointAt(t * one_over_threshold_minus_1);
                     lineToArray(pt.x(), pt.y());
